PE-4385 | cap gas limit for foundry tests (#64)
* chore: cap gas limit per test

* chore: fix breaking test due to foundry version upgrades

+++ packages/splits-v2/test/erc6909/ERC6909.t.sol
@@ -88,7 +88,7 @@ contract ERC6909Test is BaseTest {
         address _from,
         address _to,
         uint256 _id,
-        uint256 _amount
+        uint128 _amount
     )
         public
     {
@@ -96,7 +96,7 @@ contract ERC6909Test is BaseTest {
 
         vm.prank(_from);
         vm.expectRevert();
-        erc6909.transfer(_to, _id, _amount + 1);
+        erc6909.transfer(_to, _id, uint256(_amount) + 1);
     }
 
     /* -------------------------------------------------------------------------- */
